wrappers/switcherio/requirements.txt
aiohttp[speedups]
python-socketio
requests
websocket-client
